+++ src/roq/binance_futures/drop_copy.cpp
@@ -96,6 +96,9 @@ void DropCopy::operator()(const Event<Stop> &) {
 
 void DropCopy::operator()(const Event<Timer> &event) {
   connection_.refresh(event.value.now);
+  check_response_balance();
+  check_response_account();
+  check_response_orders();
 }
 
 void DropCopy::operator()(metrics::Writer &writer) {
@@ -170,6 +173,15 @@ uint32_t DropCopy::download(DropCopyState state) {
     case DropCopyState::UNDEFINED:
       assert(false);
       break;
+    case DropCopyState::BALANCE:
+      request_balance();
+      return 1;
+    case DropCopyState::ACCOUNT:
+      request_account();
+      return 1;
+    case DropCopyState::ORDERS:
+      request_orders();
+      return 1;
     case DropCopyState::DONE:
       (*this)(ConnectionStatus::READY);
       assert(!ready_);
@@ -233,7 +245,7 @@ void DropCopy::operator()(const server::Trace<json::OrderTradeUpdate> &event) {
         .last_traded_quantity = execution_report.last_filled_quantity,
         .last_traded_price = execution_report.last_filled_price,
         .last_liquidity = liquidity,
-        .update_type = {},
+        .update_type = UpdateType::INCREMENTAL,
     };
     if (shared_.update_order(
             execution_report.client_order_id,
@@ -317,6 +329,54 @@ void DropCopy::operator()(const server::Trace<json::MarginCall> &event) {
     auto &[trace_info, margin_call] = event;
     log::debug("margin_call={}"sv, margin_call);
   });
+}
+
+void DropCopy::request_balance() {
+  log::info("Requesting balance download..."sv);
+  auto &request_response = shared_.request_response[security_.get_account()];
+  request_response.request_balance = core::clock::GetSystem();
+}
+
+void DropCopy::check_response_balance() {
+  if (download_.state() != DropCopyState::BALANCE)
+    return;
+  auto &request_response = shared_.request_response[security_.get_account()];
+  if (request_response.request_balance < request_response.respond_balance) {
+    log::info("Balance download has completed!"sv);
+    download_.check(DropCopyState::BALANCE);
+  }
+}
+
+void DropCopy::request_account() {
+  log::info("Requesting account download..."sv);
+  auto &request_response = shared_.request_response[security_.get_account()];
+  request_response.request_account = core::clock::GetSystem();
+}
+
+void DropCopy::check_response_account() {
+  if (download_.state() != DropCopyState::ACCOUNT)
+    return;
+  auto &request_response = shared_.request_response[security_.get_account()];
+  if (request_response.request_account < request_response.respond_account) {
+    log::info("Account download has completed!"sv);
+    download_.check(DropCopyState::ACCOUNT);
+  }
+}
+
+void DropCopy::request_orders() {
+  log::info("Requesting order download..."sv);
+  auto &request_response = shared_.request_response[security_.get_account()];
+  request_response.request_orders = core::clock::GetSystem();
+}
+
+void DropCopy::check_response_orders() {
+  if (download_.state() != DropCopyState::ORDERS)
+    return;
+  auto &request_response = shared_.request_response[security_.get_account()];
+  if (request_response.request_orders < request_response.respond_orders) {
+    log::info("Order download has completed!"sv);
+    download_.check(DropCopyState::ORDERS);
+  }
 }
 
 }  // namespace binance_futures