+++ src/roq/binance_futures/order_entry.cpp
@@ -32,6 +32,7 @@ const auto SUPPORTS = utils::Mask{
     SupportType::CANCEL_ORDER,
     SupportType::ORDER_ACK,
     SupportType::FUNDS,
+    SupportType::POSITION,
 };
 
 struct create_metrics final : public core::metrics::Factory {
@@ -112,6 +113,24 @@ void OrderEntry::operator()(const Event<Timer> &event) {
     next_auto_cancel_ = now + Flags::rest_order_countdown() / 4;
     auto_cancel_all_open_orders();
   }
+  if (ready() && !downloading()) {
+    auto &request_response = shared_.request_response[security_.get_account()];
+    if (!downloading() && request_response.respond_balance < request_response.request_balance) {
+      log::info<1>("Download balance..."sv);
+      get_balance();
+      download_balance_ = true;
+    }
+    if (!downloading() && request_response.respond_account < request_response.request_account) {
+      log::info<1>("Download account..."sv);
+      get_account();
+      download_account_ = true;
+    }
+    if (!downloading() && request_response.respond_orders < request_response.request_orders) {
+      log::info<1>("Download orders..."sv);
+      get_open_orders();
+      download_orders_ = true;
+    }
+  }
 }
 
 void OrderEntry::operator()(metrics::Writer &writer) {
@@ -182,6 +201,9 @@ void OrderEntry::operator()(const core::web::Client::Disconnected &) {
   (*this)(ConnectionStatus::DISCONNECTED);
   if (!download_.downloading())
     download_.reset();
+  download_balance_ = false;
+  download_account_ = false;
+  download_orders_ = false;
 }
 
 void OrderEntry::operator()(const core::web::Client::Latency &latency) {
@@ -218,15 +240,6 @@ uint32_t OrderEntry::download(OrderEntryState state) {
       break;
     case OrderEntryState::LISTEN_KEY:
       get_listen_key();
-      return 1;
-    case OrderEntryState::BALANCE:
-      get_balance();
-      return 1;
-    case OrderEntryState::ACCOUNT:
-      get_account();
-      return 1;
-    case OrderEntryState::OPEN_ORDERS:
-      get_open_orders();
       return 1;
     case OrderEntryState::DONE:
       (*this)(ConnectionStatus::READY);
@@ -325,37 +338,30 @@ void OrderEntry::get_balance() {
         .body = {},
         .quality_of_service = {},
     };
-    auto sequence = download_.sequence();
-    connection_(
-        "balance"sv, request, [this, sequence]([[maybe_unused]] auto &request_id, auto &response) {
-          auto trace_info = server::create_trace_info();
-          server::Trace event(trace_info, response);
-          get_balance_ack(event, sequence);
-        });
+    connection_("balance"sv, request, [this]([[maybe_unused]] auto &request_id, auto &response) {
+      auto trace_info = server::create_trace_info();
+      server::Trace event(trace_info, response);
+      get_balance_ack(event);
+    });
   });
 }
 
-void OrderEntry::get_balance_ack(
-    const server::Trace<core::web::Response> &event, uint32_t sequence) {
+void OrderEntry::get_balance_ack(const server::Trace<core::web::Response> &event) {
   profile_.balance_ack([&]() {
     auto &[trace_info, response] = event;
-    auto state = OrderEntryState::BALANCE;
     try {
       auto [status, category, body] = response.result();
       log::debug(R"(status={}, category={}, body="{}")"sv, status, category, body);
-      if (download_.skip(sequence, state)) {
-        log::info("Download state={} has already been processed"sv, state);
-        return;
-      }
       response.expect(core::http::Status::OK);
       core::json::Buffer buffer(decode_buffer_);
       auto balance = core::json::Parser::create<json::Balance>(body, buffer);
       server::Trace event(trace_info, balance);
       (*this)(event);
-      download_.check(state);
+      download_balance_ = false;
+      auto &request_response = shared_.request_response[security_.get_account()];
+      request_response.respond_balance = core::clock::GetSystem();
     } catch (core::NetworkError &e) {
       log::warn(R"(Exception type={}, what="{}")"sv, typeid(e).name(), e.what());
-      download_.retry(state);
     }
   });
 }
@@ -396,37 +402,30 @@ void OrderEntry::get_account() {
         .body = {},
         .quality_of_service = {},
     };
-    auto sequence = download_.sequence();
-    connection_(
-        "account"sv, request, [this, sequence]([[maybe_unused]] auto &request_id, auto &response) {
-          auto trace_info = server::create_trace_info();
-          server::Trace event(trace_info, response);
-          get_account_ack(event, sequence);
-        });
+    connection_("account"sv, request, [this]([[maybe_unused]] auto &request_id, auto &response) {
+      auto trace_info = server::create_trace_info();
+      server::Trace event(trace_info, response);
+      get_account_ack(event);
+    });
   });
 }
 
-void OrderEntry::get_account_ack(
-    const server::Trace<core::web::Response> &event, uint32_t sequence) {
+void OrderEntry::get_account_ack(const server::Trace<core::web::Response> &event) {
   profile_.account_ack([&]() {
     auto &[trace_info, response] = event;
-    auto state = OrderEntryState::ACCOUNT;
     try {
       auto [status, category, body] = response.result();
       log::debug(R"(status={}, category={}, body="{}")"sv, status, category, body);
-      if (download_.skip(sequence, state)) {
-        log::info("Download state={} has already been processed"sv, state);
-        return;
-      }
       response.expect(core::http::Status::OK);
       core::json::Buffer buffer(decode_buffer_);
       auto account = core::json::Parser::create<json::Account>(body, buffer);
       server::Trace event(trace_info, account);
       (*this)(event);
-      download_.check(state);
+      download_account_ = false;
+      auto &request_response = shared_.request_response[security_.get_account()];
+      request_response.respond_account = core::clock::GetSystem();
     } catch (core::NetworkError &e) {
       log::warn(R"(Exception type={}, what="{}")"sv, typeid(e).name(), e.what());
-      download_.retry(state);
     }
   });
 }
@@ -473,39 +472,31 @@ void OrderEntry::get_open_orders() {
         .body = {},
         .quality_of_service = {},
     };
-    auto sequence = download_.sequence();
     connection_(
-        "open_orders"sv,
-        request,
-        [this, sequence]([[maybe_unused]] auto &request_id, auto &response) {
+        "open_orders"sv, request, [this]([[maybe_unused]] auto &request_id, auto &response) {
           auto trace_info = server::create_trace_info();
           server::Trace event(trace_info, response);
-          get_open_orders_ack(event, sequence);
+          get_open_orders_ack(event);
         });
   });
 }
 
-void OrderEntry::get_open_orders_ack(
-    const server::Trace<core::web::Response> &event, uint32_t sequence) {
+void OrderEntry::get_open_orders_ack(const server::Trace<core::web::Response> &event) {
   profile_.open_orders_ack([&]() {
     auto &[trace_info, response] = event;
-    auto state = OrderEntryState::OPEN_ORDERS;
     try {
       auto [status, category, body] = response.result();
       log::debug(R"(status={}, category={}, body="{}")"sv, status, category, body);
-      if (download_.skip(sequence, state)) {
-        log::info("Download state={} has already been processed"sv, state);
-        return;
-      }
       response.expect(core::http::Status::OK);
       core::json::Buffer buffer(decode_buffer_);
       auto open_orders = core::json::Parser::create<json::OpenOrders>(body, buffer);
       server::Trace event(trace_info, open_orders);
       (*this)(event);
-      download_.check(state);
+      download_orders_ = false;
+      auto &request_response = shared_.request_response[security_.get_account()];
+      request_response.respond_orders = core::clock::GetSystem();
     } catch (core::NetworkError &e) {
       log::warn(R"(Exception type={}, what="{}")"sv, typeid(e).name(), e.what());
-      download_.retry(state);
     }
   });
 }
@@ -549,7 +540,7 @@ void OrderEntry::operator()(const server::Trace<json::OpenOrders> &event) {
         .last_traded_quantity = {},
         .last_traded_price = {},
         .last_liquidity = {},
-        .update_type = {},
+        .update_type = UpdateType::SNAPSHOT,
     };
     if (shared_.update_order(
             order.client_order_id,
@@ -788,7 +779,7 @@ void OrderEntry::operator()(
       .last_traded_quantity = NaN,
       .last_traded_price = NaN,
       .last_liquidity = {},
-      .update_type = {},
+      .update_type = UpdateType::INCREMENTAL,
   };
   if (shared_.update_order(
           user_id,
@@ -983,7 +974,7 @@ void OrderEntry::operator()(
       .last_traded_quantity = NaN,
       .last_traded_price = NaN,
       .last_liquidity = {},
-      .update_type = {},
+      .update_type = UpdateType::INCREMENTAL,
   };
   if (shared_.update_order(
           user_id,