+++ src/roq/binance_futures/order_entry.h
@@ -58,6 +58,7 @@ class OrderEntry final : public core::web::Client::Handler {
   OrderEntry(const OrderEntry &) = delete;
 
   bool ready() const { return status_ == ConnectionStatus::READY; }
+  bool downloading() const { return download_balance_ || download_account_ || download_orders_; }
 
   void operator()(const Event<Start> &);
   void operator()(const Event<Stop> &);
@@ -94,15 +95,15 @@ class OrderEntry final : public core::web::Client::Handler {
   void operator()(const server::Trace<json::ListenKey> &);
 
   void get_balance();
-  void get_balance_ack(const server::Trace<core::web::Response> &, uint32_t sequence);
+  void get_balance_ack(const server::Trace<core::web::Response> &);
   void operator()(const server::Trace<json::Balance> &);
 
   void get_account();
-  void get_account_ack(const server::Trace<core::web::Response> &, uint32_t sequence);
+  void get_account_ack(const server::Trace<core::web::Response> &);
   void operator()(const server::Trace<json::Account> &);
 
   void get_open_orders();
-  void get_open_orders_ack(const server::Trace<core::web::Response> &, uint32_t sequence);
+  void get_open_orders_ack(const server::Trace<core::web::Response> &);
   void operator()(const server::Trace<json::OpenOrders> &);
 
   void refresh_listen_key();
@@ -179,6 +180,9 @@ class OrderEntry final : public core::web::Client::Handler {
   // experimental
   absl::flat_hash_set<std::string> open_orders_symbols_;
   std::chrono::nanoseconds next_auto_cancel_ = {};
+  bool download_balance_ = false;
+  bool download_account_ = false;
+  bool download_orders_ = false;
 };
 
 }  // namespace binance_futures